+++ src/Serilog.Sinks.Http/Sinks/Http/Private/Time/ExponentialBackoffConnectionSchedule.cs
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace Serilog.Sinks.Http.Private
+namespace Serilog.Sinks.Http.Private.Time
 {
 	internal class ExponentialBackoffConnectionSchedule
 	{