+++ test/Serilog.Sinks.Http.IntegrationTests.Server/Controllers/Dtos/EventDto.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Serilog.Sinks.Http.IntegrationTests.Server.Controllers
+namespace Serilog.Sinks.Http.IntegrationTests.Server.Controllers.Dtos
 {
 	public class EventDto
 	{
@@ -11,8 +11,12 @@ namespace Serilog.Sinks.Http.IntegrationTests.Server.Controllers
 
 		public string MessageTemplate { get; set; }
 
-		public Dictionary<string, string> Properties { get; set; }
+		public string RenderedMessage { get; set; }
 
 		public string Exception { get; set; }
+
+		public Dictionary<string, string> Properties { get; set; }
+
+		public Dictionary<string, RenderingDto[]> Renderings { get; set; }
 	}
 }